nulle w srodku wierszy

+++ parser.h
@@ -1,6 +1,6 @@
 #ifndef PARSER_H
 #define PARSER_H
 
-extern int process_line(char* line, long double **n, int *n_i, char ***nan, int *nan_i, size_t read);
+extern int process_line(char* line, long double **n, int *n_i, char ***nan, int *nan_i, int read);
 
 #endif /* PARSER_H */
